feat: add optionalResolverInfo option.
This feature adds the option to generate resolver types with an
optional info argument (the last argument of a resolver).

This option is useful for projects that do not make use of the info
argument and want to simplifier their resolvers and their testing by
omitting the parameter from their resolver function definitions.